Rename `api_testing_helpers` to `api_testing_common`
