+++ concent_api/api-integration-force-payment.py
@@ -12,8 +12,8 @@ from golem_messages         import message
 from utils.helpers          import get_current_utc_timestamp
 from utils.testing_helpers  import generate_ecc_key_pair
 
-from api_testing_helpers    import api_request
-from api_testing_helpers    import timestamp_to_isoformat
+from api_testing_common     import api_request
+from api_testing_common     import timestamp_to_isoformat
 
 import requests
